refactor(sentry): split the two modules this PR pushed over the 600-line cap
Measured against main, this stack grew two files past the soft cap. Split both
along their natural seam; no behaviour change, and no new test file (the tests
stay in the suites that already cover these modules, so the suite-registration
cascade is not triggered).

scripts/sentry-autofix-select.mjs (main 565 -> 662 -> 442)
  Extracted the per-stub eligibility layer into
  scripts/sentry-autofix-candidate.mjs (248): evaluateCandidate and safeShortId,
  plus the AUTOFIX_VERDICT / SKIP_FIX_SCOPE_ARCHITECTURAL constants they own.
  That is every filter deciding whether ONE stub may start a fix attempt — the
  terminal markers, the queue label, the SHORT-ID parse, the verdict re-parse
  and its exactly-local repo check, the reconcile branch and the fix_scope gate.
  select.mjs keeps the window listing, the family collapse and its CLI, so the
  workflow entrypoint and every export the suites import are unchanged
  (SKIP_FIX_SCOPE_ARCHITECTURAL is re-exported).

scripts/sentry-autofix-record-labels.mjs (605 -> 643 -> 441)
  Extracted the revalidation/compensation layer into
  scripts/sentry-autofix-hold-revalidate.mjs (242): liveArchitecturalScope (the
  live re-read bracketing the write) and withdrawStaleHold (the remove plus the
  terminal-guarded re-queue), with readStubWithState and the bookkeeping note.
  record-labels keeps the pure plan, the label write loop and its CLI.

scripts/sentry-triage-digest.mjs is deliberately NOT split: main 933 -> 699 is a
net reduction of 234 lines from the digest-render extraction, the same
drift-down argument accepted for finalize.mjs.

Gate routing follows the modules: candidate.mjs joins the selection-helper case
(select:test), hold-revalidate.mjs joins record-labels (finalize:test). Both
routes verified by a dry-run against the two paths. The suite-gate integrity
watch set is derived from static imports, so it picks up both new modules with
no manual registration.
